Cleaned up macros and switched to 32-bit move command

+++ shared_linux/TestStandCommHost/TestStandCommHost.h
@@ -11,7 +11,7 @@ class TestStandCommHost : public TestStandComm
         bool ping();
         bool get_status();
         bool home();
-        bool move(uint16_t accel, uint16_t hold_vel, uint16_t dist, AxisId axis, DirectionId dir);
+        bool move(uint32_t accel, uint32_t hold_vel, uint32_t dist, AxisId axis, DirectionId dir);
         bool stop();
         bool get_data(DataId data_id);
 };
